chore: Move QueryMap to its own file
